+++ doc/mysql-schema/frostmourne.sql
@@ -369,6 +369,37 @@ CREATE TABLE IF NOT EXISTS `service_info`
 ) ENGINE = InnoDB
   DEFAULT CHARSET = utf8mb4 COMMENT ='服务信息';
 
+/*------------------------------------------- create short_link---------------------------------------------------------------------*/
+DROP TABLE IF EXISTS short_link;
+CREATE TABLE IF NOT EXISTS short_link
+(
+    id        BIGINT AUTO_INCREMENT PRIMARY KEY COMMENT '自增主键',
+    long_link VARCHAR(1000) NOT NULL COMMENT '监控数据类型。(http: http监控；其他值: 关联data_name表)',
+    create_at DATETIME      NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT '创建时间'
+)
+    ENGINE = InnoDB
+    DEFAULT CHARSET = utf8mb4
+    COMMENT = '短链接表';
+
+/*------------------------------------------- create config_map---------------------------------------------------------------------*/
+DROP TABLE IF EXISTS config_map;
+CREATE TABLE IF NOT EXISTS config_map
+(
+    id           BIGINT AUTO_INCREMENT PRIMARY KEY COMMENT '自增主键',
+    config_key   VARCHAR(500)  NOT NULL COMMENT '配置key',
+    config_value VARCHAR(1000) NOT NULL COMMENT '配置value',
+    creator      VARCHAR(200)  NOT NULL COMMENT '创建人',
+    create_at    DATETIME      NOT NULL COMMENT '创建时间',
+    modify_at    DATETIME      NOT NULL COMMENT '修改时间',
+    modifier     VARCHAR(200)  NOT NULL COMMENT '最后修改人'
+)
+    ENGINE = InnoDB
+    DEFAULT CHARSET = utf8mb4
+    COMMENT = '内部配置表';
+
+CREATE UNIQUE INDEX uniq_configkey ON config_map (config_key);
+
+
 /*------------------------------------------- init data---------------------------------------------------------------------*/
 INSERT INTO department_info(department_name, full_name, creator, create_at, modify_at, modifier)
 VALUES ('default', '默认部门', 'admin', now(), now(), 'admin');